requirements.txt
streamlit
pandas
numpy
scipy
plotly
pillow
matplotlib
scikit-learn
logisticR.py
datetime
seaborn
statsmodels.api
